foo.js
Node.js has a simple module loading system. In Node.js, files and modules are in one-to-one correspondence. As an example, foo.js loads the module circle.js in the same directory.To build the addon you have to use the addon builder sdk from mozilla.

The contents of foo.js: